Fixing SRA tests
